+++ tests/fixtures/plugins/prettier.config.js
@@ -1,3 +1,3 @@
 module.exports = {
-  plugins: ['../../..'],
+  plugins: ['../../../dist/index.js'],
 }
