+++ tests/utils.js
@@ -1,3 +1,6 @@
+const prettier = require('prettier')
+const path = require('path')
+
 let testClassName = 'sm:p-0 p-0'
 let testClassNameSorted = 'p-0 sm:p-0'
 let yes = '__YES__'
@@ -21,4 +24,21 @@ module.exports.t = function t(strings, ...values) {
   })
 
   return [input, output]
+}
+
+module.exports.format = async function format(str, options = {}) {
+  let result = await prettier.format(str, {
+    pluginSearchDirs: [__dirname], // disable plugin autoload
+    semi: false,
+    singleQuote: true,
+    printWidth: 9999,
+    parser: 'html',
+    ...options,
+    plugins: [
+      ...options.plugins ?? [],
+      path.resolve(__dirname, '../dist/index.js')
+    ],
+  })
+
+  return result.trim()
 }
